+++ src/cqfantasy/tower/BaseSection.py
@@ -1,0 +1,5 @@
+import cadquery as cq
+from cadqueryhelper import Base
+
+class BaseSection(Base):
+    pass