+++ src/cqfantasy/tower/__init__.py
@@ -1,11 +1,19 @@
 
+try:
+    log #type:ignore
+except NameError:
+    log = print
+
 from .cut_cylinder import cut_cylinder
+
 from .magnets import make_magnet, make_magnets
 from .TowerWindow import TowerWindow
 from .TowerDoor import TowerDoor
 from .FrameWindow import FrameWindow
 from .LatticeWindow import LatticeWindow
 from .TileGenerator import TileGenerator
+from .RoundBlockGenerator import RoundBlockGenerator
+from .BaseSection import BaseSection
 from .TowerBase import TowerBase
 from .UnevenBlocks import UnevenBlocks
 from .TowerBaseGreebled import TowerBaseGreebled